Move to FFI
It was simply too fiddly to be allocating memory for strings
and was only going to get worse with typed lists in the future.

So now I'm using FFI to abstract the boring pointer math away.
Fiddle was cool but limited.

This also implements the first string function, #reverse. \o/